+++ benchmarks/verus/apps/files/apps_test_4681.rs
@@ -25,9 +25,9 @@ spec fn valid_input(n: int) -> bool {
 // </vc-helpers>
 
 // <vc-spec>
-fn solve(n: int) -> (result: int)
-    requires valid_input(n)
-    ensures result == lucas(n)
+fn solve(n: i8) -> (result: i8)
+    requires valid_input(n as int)
+    ensures result as int == lucas(n as int)
 // </vc-spec>
 // <vc-code>
 {
