+++ benchmarks/verus/apps/files/apps_test_811.rs
@@ -25,12 +25,12 @@ proof fn a_decreases(a: int, b: int) {
 // </vc-helpers>
 
 // <vc-spec>
-fn solve(a: int, b: int) -> (result: int)
+fn solve(a: i8, b: i8) -> (result: i8)
   requires 
-    valid_input(a, b)
+    valid_input(a as int, b as int)
   ensures 
-    result >= a,
-    result == total_burning_hours(a, b)
+    result as int >= a as int,
+    result as int == total_burning_hours(a as int, b as int)
 // </vc-spec>
 // <vc-code>
 {
